Ensure plot defs are always created
When the Go process stops, the web UI tries to reconnect to the
websocket endpoint periodically. When we start a new Go process
'plotdefs' won't be called since the web UI already has all
files. On the Go side, we initialize the plot definitions when we
establish the wsebsocket connection.

We protect plotdefs with sync.Once to avoid concurrent accesses.