Fix testWs
Before, the data sent via websocket was static. Now that the set of
plots is dynamic, it would be overkill to exhaustively check all time
series and heatmaps we send, plus that varies from one go version to
another.

Instead, check that we receive 2 data points which presence are likely
guarantee:
 - a time series: the number of goroutines
 - the 'size classes' heatmap