internal/plot: refactor/simplify plot description
Handle metric indices on a per go version basis (for now only
pre and post 1.26) and declare them as globals, thus allowing plot
description from not requiring to store them.

We now also don't need to have an object, since we need only need a
very thing state, if at all, to compute plot values from metrics. In
fact we just need a function. So declare the getvalues closure directly
in the declaration of each plot description{}.

With the additions of small wrappers to compute delta and rate from
metrics, the getvalues becomes very short and simple, much more maintable.