internal/plots: refactor + add new go1.26 plots (#142)
* internal/plot: use type for tags

* _example: slight change in Work

* Add new /sched/goroutines metrics to goroutines plot

* Add new threads plot

* Use cgo time series in test

* internal/plot: move each plot in its own file

* internal/plot: each plot registers itself

* internal/plot: use rgb color comments for heatmap shades

* internal/plot: fixes and cosmetics

* internal/plot: refactor/simplify plot description

Handle metric indices on a per go version basis (for now only
pre and post 1.26) and declare them as globals, thus allowing plot
description from not requiring to store them.

We now also don't need to have an object, since we need only need a
very thing state, if at all, to compute plot values from metrics. In
fact we just need a function. So declare the getvalues closure directly
in the declaration of each plot description{}.

With the additions of small wrappers to compute delta and rate from
metrics, the getvalues becomes very short and simple, much more maintable.

* internal/plot: add prego1.26 indices file

* internal/plot: generate indices

* CONTRIBUTING.md: fix command 'npm run dev'

* internal/plot: convert more plots

* internal/plot: remove sched-events plot

* internal/plot: all plots converted

* internal/plot: use sync.Once for initIndices() to ensure init() order

* internal/plots: remove godebug indices

* ci: on linux, test on 3 latest Go versions

* ci: run less often

* _example: remove go.sum

* _example: add go.sum

* internal/plot: simplification

* internal/plot: simplify description

* Cosmetics

* internal/plot: end refactor registry and List

* internal/plot: rewrite TestUnusedRuntimeMetrics

* internal/plot: determine order

By calling mustidx in the global scope rather than in init(), we
ensure that metric indices are known when registering plots, which help
for dynamic layout building required for histograms.

* internal/plot: move stuff in helpers.go

* internal/plot: build histogram layout at "register time"

* ci: run linux tests on 1.24, 1.25 and tip

* Support multiple websocket clients. Add Server.Close()

* Add TestWsConcurrent

* Fix test. Init client in Server.init()

* fixup! Support multiple websocket clients. Add Server.Close()

* Cosmetics

* Fix goroutine leak. Add synctest test

* Cosmetics

* Cosmetics

* Accept Root("/") option

* internal/plot: deterministic order in TestUnused